backend/requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.0.7
django-common-helpers==0.9.2
django-cors-headers==3.2.0
django-cors-middleware==1.5.0
django-cron==0.5.1
django-enumchoicefield==2.0.0
django-heroku==0.3.1
django-oauth-toolkit==1.3.2
django-oauth2-provider==0.2.6.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
oauthlib==3.1.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
python-dotenv==0.14.0
pytz==2019.3
requests==2.24.0
shortuuid==1.0.1
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.9
wrapt==1.11.2
